Add a file upload option for template
Add a field for file upload and allow the uploaded file be saved and used
in place of the default templates. This is an intial implementation of
this feature. More work will be done to integrate the upload field with
the existing template field. Will also implement safety features such as
max file size and type.